Move package verifier into automation module
