+++ scripts/package_verify.sh
@@ -3,7 +3,8 @@ set -e
 export PYTHONPATH=
 virtualenv package-verify-env
 . package-verify-env/bin/activate
-python scripts/command_modules/package_verify.py
+pip install -e scripts
+python -m automation.tests.verify_packages
 deactivate
 # If we get here, all prev. commands returned 0 exit codes so we are done.
 rm -rf package-verify-env
